+++ src/Inspectors/Instance/InstanceInspector.h
@@ -1,7 +1,7 @@
 #ifndef REALMOPS_INSTANCE_INSPECTOR_H
 #define REALMOPS_INSTANCE_INSPECTOR_H
 
-#include "CommandScript.h"
+#include "Chat.h"
 #include "Define.h"
 
 class ChatHandler;
@@ -12,7 +12,7 @@ class InstanceInspector
 {
 public:
     static bool Search(ChatHandler* handler, Acore::ChatCommands::Tail search);
-    static bool Audit(ChatHandler* handler, uint32 mapId, Optional<uint8> difficultyArg);
+    static bool Audit(ChatHandler* handler, uint32 mapId, Acore::ChatCommands::Optional<uint8> difficultyArg);
 };
 }